[feat]: navigate confirm place

+++ data/exhibition/src/main/java/com/example/exhibition/model/remote/request/RequestPostPlaceDto.kt
@@ -10,9 +10,9 @@ data class RequestPostPlaceDto(
     @SerialName("id")
     val id: String,
     @SerialName("latitude")
-    val latitude: Int,
+    val latitude: Double,
     @SerialName("longitude")
-    val longitude: Int,
+    val longitude: Double,
     @SerialName("name")
     val name: String,
 )
